Implemented login and logout functionality.
This branch introduces user authentication and access control, focusing on secure login, logout, and authorization for room management. It restricts access to specific pages and actions based on user authentication status and ownership.

+++ studyhub/base/views.py
@@ -1,11 +1,43 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
 
 # Create your views here.
 
 
+
+def loginPage(request): # Don't call the login view Login(), that is a Django function and will cause conflict
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist.')
+        
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else: 
+            messages.error(request, 'Username or password incorrect.')
+    context = {}
+    return render(request, 'base/login-register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -24,6 +56,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm
     if request.method == 'POST':
@@ -34,9 +67,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room-form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room) # The form will be prefilled with the room value
+
+    if request.user != room.host:
+        return HttpResponse("You don't have permission to carry out that action")
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -45,8 +83,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room-form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id= pk)
+
+    if request.user != room.host:
+        return HttpResponse("You don't have permission to carry out that action")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
